+++ packages/xgplayer-hls/src/hls/config.js
@@ -15,6 +15,9 @@
  *  loadTimeout?: number,
  *  preloadTime?: number,
  *  disconnectTime?: number,
+ *  allowedStreamTrackChange?: boolean,
+ *  seiInTime?: boolean,
+ *  manifestList?: Array<{url: string, manifest: string}>
  *  fetchOptions?: RequestInit
  *  onPreM3U8Parse?: (m3u8: string) => string | void
  *  decryptor?: Decryptor
@@ -41,6 +44,8 @@ export function getConfig (cfg) {
     targetLatency: 10,
     maxLatency: 20,
     allowedStreamTrackChange: true,
+    seiInTime: false,
+    manifestList: [],
     ...cfg,
     media
   }